fix: html of options js distinguise
