+++ js/index.js
@@ -6,3 +6,4 @@ import * as charts from './charts';
 import * as maps from './maps';
 import * as custom from './custom';
 import * as dateRange  from './date-range';
+import * as layoutSettings  from './layout-settings';
