+++ scss/sleek.scss
@@ -42,5 +42,6 @@
 @import 'widgets';
 @import 'error';
 @import 'lada-button';
+@import 'optionswitch';
 @import 'github';
 @import 'rtl';
