+++ src/_includes/base/head.html
@@ -37,11 +37,6 @@
   <!-- SLEEK CSS -->
   <link id="sleek-css" rel="stylesheet" href="{% if page.rtl == true %}assets/css/sleek.rtl.css{% else %}assets/css/sleek.css{% endif %}" />
 
-  {% if site.environment == "demo" %}
-  <!-- CSS for Demo -->
-  <link rel="stylesheet" href="assets/options/optionswitch.css" />
-  {% endif %}
-
   <!-- FAVICON -->
   <link href="assets/img/favicon.png" rel="shortcut icon" />
