+++ src/_includes/base/header.html
@@ -90,8 +90,8 @@
                       <li>
                         <a href="#"> <i class="mdi mdi-diamond-stone"></i> Projects </a>
                       </li>
-                      <li>
-                        <a href="#"> <i class="mdi mdi-settings"></i> Account Setting </a>
+                      <li class="">
+                        <a href="#"> <i class="mdi mdi-settings"></i> Setting </a>
                       </li>
 
                       <li class="dropdown-footer">
