+++ src/_includes/base/scripts.html
@@ -1,10 +1,4 @@
 <script src="assets/plugins/jquery/jquery.min.js"></script>
-
-{% if site.environment == "demo" %}
-<!-- JS for Demo -->
-<script src="assets/options/optionswitcher.js"></script>
-{% endif %}
-
 <script src="assets/plugins/slimscrollbar/jquery.slimscroll.min.js"></script>
 <script src="assets/plugins/jekyll-search.min.js"></script>
