+++ src/_layouts/default.htm
@@ -44,7 +44,7 @@
         {% include base/right-sidebar.html %}
         {% endif %}
 
-
+        {% include base/layout-settings.html %}
       </div>
 
       {% include base/footer.html %}
